refactor old codes - optimize code - delete unnecessary classes- fix buy sell button issue
