--- app/src/main/java/com/finance/trade_learn/models/DataForHomePage.kt
@@ -1,8 +1,0 @@
-package com.finance.trade_learn.models
-
-import com.finance.trade_learn.models.modelsConvector.CoinsHome
-
-class DataForHomePage (
-    val lastCrypoList: List<CoinsHome>,
-    var ListOfCrypto : ArrayList<CoinsHome>,
-    )
