--- app/src/main/java/com/finance/trade_learn/models/enumPriceChange.kt
@@ -1,6 +1,0 @@
-package com.finance.trade_learn.models
-
-enum class enumPriceChange {
-    pozitive,negative,notr
-
-}
