+++ app/src/main/java/com/finance/trade_learn/models/modelsConvector/CoinsHome.kt
@@ -1,7 +1,5 @@
 package com.finance.trade_learn.models.modelsConvector
 
-import com.finance.trade_learn.models.enumPriceChange
-
 data class CoinsHome(
     val id : String,
     val CoinName: String,
@@ -9,7 +7,6 @@ data class CoinsHome(
     val CoinPrice: String,
     val CoinChangePercente: String,
     val CoinImage: String,
-    val raise: enumPriceChange,
     val marketCap : String,
     val total_volume : String,
 )
